pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cinema-summary-bot"
version = "0.1.0"
description = "Telegram bot MVP for spoiler-aware, source-grounded movie explanations"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "gigachat>=0.2.0",
  "uvicorn>=0.30.0",
  "httpx>=0.27.0",
  "pydantic>=2.8.0",
  "python-telegram-bot>=22.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
